Oly empty SVBTree can be removed to make deletion of indexes durable operation

+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/singlevalue/OCellBTreeSingleValue.java
@@ -402,6 +402,9 @@ public final class OCellBTreeSingleValue<K> extends ODurableComponent {
     try {
       acquireExclusiveLock();
       try {
+        if (size() > 0) {
+          throw new OCellBTreeSingleValueException("Only empty tries can be removed", this);
+        }
         deleteFile(fileId);
         deleteFile(nullBucketFileId);
       } finally {
